Refatoração do Gerenciador de Pessoas ('PessoaManager') para que atenda ao padrão definido no Gerenciador de Contratos ('ContratoEmprestimoManager'), passando de interface para classe abstrata, refatorando também as classes dependentes 'Impl' e 'Mock'.
