+++ src/main/java/br/com/cresol/desafio/manager/impl/PessoaManagerImpl.java
@@ -1,0 +1,21 @@
+package br.com.cresol.desafio.manager.impl;
+
+import java.util.Map;
+
+import br.com.cresol.desafio.dto.Pessoa;
+import br.com.cresol.desafio.manager.PessoaManager;
+
+public class PessoaManagerImpl extends PessoaManager {
+
+	@Override
+	protected Map<String, Pessoa> recuperaPessoasPersistidas() {
+		// TODO-REVER - Implementar uma persistencia (ex.: diretório "standalone/data")
+		return null;
+	}
+
+	@Override
+	protected void persisteAdicaoDePessoa(Pessoa pessoa) {
+		// TODO-REVER - Implementar uma persistencia (ex.: diretório "standalone/data")
+	}
+
+}
